Feature/align with bowerbird test updates

+++ test/bowerbird-deps/test-git-dependency.mk
@@ -1,11 +1,13 @@
 test-git-dependency-bowerbird-deps-success:
+	test ! -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps || rm -rf $(WORKDIR_TEST)/$@/deps/bowerbird-deps
+	test ! -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 	$(MAKE) FORCE TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_SUCCESS=true
 	test -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 	! test -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps/.git
 
 ifdef TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_SUCCESS
-    $(eval $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-success/deps/bowerbird-deps,\
-			https://github.com/ic-designer/make-bowerbird-deps.git,main,bowerbird.mk))
+    $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-success/deps/bowerbird-deps,\
+			https://github.com/ic-designer/make-bowerbird-deps.git,main,bowerbird.mk)
 endif
 
 
@@ -14,19 +16,19 @@ test-git-dependency-bowerbird-deps-bad-entry:
 	! test -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 
 ifdef TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_BAD_ENTRY
-    $(eval $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-entry/deps/bowerbird-deps,\
-			https://github.com/ic-designer/make-bowerbird-deps.git,main,))
+    $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-entry/deps/bowerbird-deps,\
+			https://github.com/ic-designer/make-bowerbird-deps.git,main,)
 endif
 
 
 test-git-dependency-bowerbird-deps-bad-path:
-	rm -rf $(WORKDIR_TEST)/$@/deps/bowerbird-deps
+	test ! -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps || rm -rf $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 	mkdir -p $(WORKDIR_TEST)/$@/deps/bowerbird-deps/.git
 	! $(MAKE) FORCE TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_BAD_PATH=true
 
 ifdef TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_BAD_PATH
-    $(eval $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-path/deps/bowerbird-deps,\
-			https://github.com/ic-designer/make-bowerbird-deps.git,main,bowerbird.mk))
+    $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-path/deps/bowerbird-deps,\
+			https://github.com/ic-designer/make-bowerbird-deps.git,main,bowerbird.mk)
 endif
 
 
@@ -35,8 +37,8 @@ test-git-dependency-bowerbird-deps-bad-url:
 	! test -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 
 ifdef TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_BAD_URL
-    $(eval $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-url/deps/bowerbird-deps,\
-			,,main,bowerbird.mk))
+    $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-url/deps/bowerbird-deps,\
+			,,main,bowerbird.mk)
 endif
 
 
@@ -45,8 +47,8 @@ test-git-dependency-bowerbird-deps-bad-version:
 	! test -d $(WORKDIR_TEST)/$@/deps/bowerbird-deps
 
 ifdef TEST_GIT_DEPENDENCY_BOWERBIRD_DEPS_BAD_VERSION
-        $(eval $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-version/deps/bowerbird-deps,\
-			https://github.com/ic-designer/make-bowerbird-deps.git,,bowerbird.mk))
+        $(call bowerbird::git-dependency,$(WORKDIR_TEST)/test-git-dependency-bowerbird-deps-bad-version/deps/bowerbird-deps,\
+			https://github.com/ic-designer/make-bowerbird-deps.git,,bowerbird.mk)
 endif
